Move TypeInfo's DT generation into todt.c
